chore(scanner): followup to #6663
