+++ scanner/cmd/scannerctl/main.go
