+++ scanner/go.mod
@@ -9,6 +9,7 @@ require (
 	github.com/quay/zlog v1.1.4
 	github.com/rs/zerolog v1.29.1
 	github.com/stackrox/rox v0.0.0
+	golang.org/x/sys v0.10.0
 	google.golang.org/grpc v1.56.2
 )
 
@@ -96,7 +97,6 @@ require (
 	golang.org/x/net v0.12.0 // indirect
 	golang.org/x/oauth2 v0.10.0 // indirect
 	golang.org/x/sync v0.3.0 // indirect
-	golang.org/x/sys v0.10.0 // indirect
 	golang.org/x/term v0.10.0 // indirect
 	golang.org/x/text v0.11.0 // indirect
 	golang.org/x/time v0.3.0 // indirect
